mux: remove redundant cast
This should have been part of
    2019-04-29 Use better command-line args for lbs, mux, s3, lbs-s3
    22d5504a2805e94a452e32b5c1297b975c446ca5
when we changed opt_n from intmax_t to size_t.

Reported by:	clang-tidy

+++ mux/main.c
@@ -178,7 +178,7 @@ main(int argc, char * argv[])
 
 	/* Initialize the dispatcher. */
 	if ((dstate = dispatch_init(socks_s, opt_s_size,
-	    Q_t, opt_n ? (size_t)opt_n : SIZE_MAX)) == NULL) {
+	    Q_t, opt_n ? opt_n : SIZE_MAX)) == NULL) {
 		warnp("Failed to initialize dispatcher");
 		exit(1);
 	}
